# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Karma Yogi API
  version: 1.0.0
  description: |
    Public HTTP API for Karma Yogi. All paths below are rooted at `/api/v1` (e.g. full URL `http://localhost:8080/api/v1/auth/login`).
    Authenticated routes expect `Authorization: Bearer <accessToken>`.
servers:
  - url: /api/v1
paths:
  /healthz:
    get:
      tags: [System]
      summary: Health check
      description: Service-level health endpoint exposed at server root (not under `/api/v1`).
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
  /auth/google:
    post:
      tags: [Auth]
      summary: Exchange Google ID token for app session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: Google ID token (JWT) from Google Identity Services
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid or unverifiable Google token
  /auth/dev-login:
    post:
      tags: [Auth]
      summary: Development convenience login (creates or reuses a user)
      description: |
        Optional body `{"email":"..."}`. If `email` is omitted or empty, defaults to `dev@karmayogi.local`.
        Returns the same shape as other auth flows. Prefer email/password or Google in real environments.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Bad request
  /auth/register:
    post:
      tags: [Auth]
      summary: Register with email, password, and security answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, secretAnswer]
              properties:
                email:
                  type: string
                  format: email
                fullName:
                  type: string
                password:
                  type: string
                  minLength: 8
                secretAnswer:
                  type: string
                  description: Answer to the standard security question (client displays the prompt)
      responses:
        "200":
          description: Registered and authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Validation or registration error
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials or password login not available for this account
  /auth/password-reset:
    post:
      tags: [Auth]
      summary: Reset password using email and security answer (no email delivery)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, secretAnswer, newPassword]
              properties:
                email:
                  type: string
                  format: email
                secretAnswer:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "400":
          description: |
            Unable to reset. JSON body uses `error` string. Common values:
            - `incorrect email or security answer` (includes unknown email to limit enumeration)
            - `this account uses google sign-in; password reset is not available`
            - `this account does not have a security answer on file; create a new account or contact support`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshId, refreshToken]
              properties:
                refreshId:
                  type: string
                refreshToken:
                  type: string
      responses:
        "200":
          description: New tokens issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid or expired refresh session
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshId]
              properties:
                refreshId:
                  type: string
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "400":
          description: Bad request
  /users/me:
    get:
      tags: [Users]
      summary: Current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
    patch:
      tags: [Users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Validation error
  /subjects:
    get:
      tags: [Subjects]
      summary: List subjects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Subjects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subject"
        "400":
          description: Bad request
    post:
      tags: [Subjects]
      summary: Create subject
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubjectCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subject"
        "400":
          description: Bad request
  /subjects/{id}:
    delete:
      tags: [Subjects]
      summary: Delete subject (sessions under it are removed)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "404":
          description: Not found
  /sessions:
    get:
      tags: [Sessions]
      summary: List sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
        "400":
          description: Bad request
    post:
      tags: [Sessions]
      summary: Create session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Bad request
  /sessions/{id}:
    patch:
      tags: [Sessions]
      summary: Update session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionUpdateRequest"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Bad request
    delete:
      tags: [Sessions]
      summary: Delete session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusDeleted"
        "404":
          description: Not found
  /goals:
    get:
      tags: [Goals]
      summary: List goals
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
        "400":
          description: Bad request
    post:
      tags: [Goals]
      summary: Create goal
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "400":
          description: Bad request
  /goals/{id}:
    patch:
      tags: [Goals]
      summary: Update goal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdateRequest"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "400":
          description: Bad request
    delete:
      tags: [Goals]
      summary: Delete goal
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusDeleted"
        "404":
          description: Not found
  /insights:
    get:
      tags: [Insights]
      summary: Aggregated insights for the current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Insights snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insights"
        "400":
          description: Bad request
  /timer-state:
    get:
      tags: [Timer]
      summary: Get persisted timer state for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Timer state payload or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimerStateResponse"
        "400":
          description: Bad request
    put:
      tags: [Timer]
      summary: Upsert persisted timer state for current user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state]
              properties:
                state:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "400":
          description: Bad request
    delete:
      tags: [Timer]
      summary: Clear persisted timer state for current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatusOk"
        "400":
          description: Bad request
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    StatusOk:
      type: object
      properties:
        status:
          type: string
          example: ok
    StatusDeleted:
      type: object
      properties:
        status:
          type: string
          example: deleted
    ErrorBody:
      type: object
      properties:
        error:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        username:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateUserRequest:
      type: object
      properties:
        fullName:
          type: string
        username:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
    AuthResult:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        accessToken:
          type: string
        refreshId:
          type: string
        refreshToken:
          type: string
    Subject:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        name:
          type: string
        color:
          type: string
        icon:
          type: string
        createdAt:
          type: string
          format: date-time
    SubjectCreateRequest:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
        color:
          type: string
    Session:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        subjectId:
          type: string
        topic:
          type: string
        durationMin:
          type: integer
        mood:
          type: string
        startedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    SessionCreateRequest:
      type: object
      required: [subjectId, topic, durationMin, mood, startedAt]
      properties:
        subjectId:
          type: string
        topic:
          type: string
        durationMin:
          type: integer
        mood:
          type: string
        startedAt:
          type: string
          format: date-time
    SessionUpdateRequest:
      type: object
      required: [topic, durationMin, mood, startedAt]
      properties:
        topic:
          type: string
        durationMin:
          type: integer
        mood:
          type: string
        startedAt:
          type: string
          format: date-time
    Goal:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        title:
          type: string
        targetMinutes:
          type: integer
        deadline:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GoalCreateRequest:
      type: object
      required: [title, targetMinutes, deadline]
      properties:
        title:
          type: string
        targetMinutes:
          type: integer
        deadline:
          type: string
          format: date-time
    GoalUpdateRequest:
      type: object
      required: [title, targetMinutes, deadline]
      properties:
        title:
          type: string
        targetMinutes:
          type: integer
        deadline:
          type: string
          format: date-time
    Insights:
      type: object
      properties:
        totalMinutes:
          type: integer
        sessionCount:
          type: integer
        weeklyMinutes:
          type: integer
        goalCompletion:
          type: integer
    TimerStateResponse:
      type: object
      properties:
        state:
          type: object
          nullable: true
          additionalProperties: true